data plane initial integration

+++ src/app/ServiceRegistry/ServiceRegistry.tsx
@@ -1,30 +1,30 @@
 import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
-import { PageSection, ButtonVariant } from '@patternfly/react-core';
-import { MASEmptyState, MASEmptyStateVariant } from '@app/components';
+import { DefaultApi, Configuration, Registry } from '@rhoas/registry-management-sdk';
+import { useAuth, useConfig } from '@bf2/ui-shared';
 import {
-  ServiceRegistryHeader,
   ServiceRegistryDrawer,
   FederatedModuleManager,
   FederatedModuleManagerProps,
+  UnauthrizedUser,
+  WelcomeEmptyState,
 } from './components';
-import { DefaultApi, Configuration } from '@rhoas/registry-management-sdk';
-import { useAuth } from '@bf2/ui-shared';
+import { ServiceRegistryHeader } from '@app/ServiceRegistry/components';
 
-export type ServiceRegistryProps = Pick<
-  FederatedModuleManagerProps,
-  'params' | 'homeLinkPath' | 'baseUIPath' | 'navPrefixPath'
->;
+export type ServiceRegistryProps = FederatedModuleManagerProps;
 
 export const ServiceRegistry: React.FC<ServiceRegistryProps> = ({
-  navPrefixPath,
   baseUIPath,
   params,
   homeLinkPath,
+  showBreadcrumb,
+  activeFederatedModule,
+  activeBreadcrumbItemLabel,
 }) => {
-  const { t } = useTranslation();
   const auth = useAuth();
+  // const {
+  //   srs: { apiBasePath: basePath },
+  // } = useConfig();
   const history = useHistory();
 
   const [isExpandedDrawer, setIsExpandedDrawer] = useState<boolean>(false);
@@ -32,8 +32,8 @@ export const ServiceRegistry: React.FC<ServiceRegistryProps> = ({
   const [serviceAccountDetails, setServiceAccountDetails] = useState<any>(undefined);
   const [notRequiredDrawerContentBackground, setNotRequiredDrawerContentBackground] = useState<boolean>(false);
   const [isUnauthorizedUser, setIsUnauthorizedUser] = useState<boolean>(false);
-  const [registry, setRegistry] = useState();
-  const basePath = 'http://localhost:9090';
+  const [registry, setRegistry] = useState<Registry>();
+  const basePath = 'https://api.stage.openshift.com';
   const { name: tenantId } = registry || {};
   const registryId = 'tenant-15';
 
@@ -88,18 +88,20 @@ export const ServiceRegistry: React.FC<ServiceRegistryProps> = ({
 
   const deleteRegistry = async () => {
     const accessToken = await auth?.kas.getToken();
-    const api = new DefaultApi(
-      new Configuration({
-        accessToken,
-        basePath,
-      })
-    );
-    try {
-      await api.deleteRegistry(registry?.id).then(() => {
-        history.push(homeLinkPath || '/');
-        fetchRegistry();
-      });
-    } catch (error) {}
+    if (registry?.id) {
+      const api = new DefaultApi(
+        new Configuration({
+          accessToken,
+          basePath,
+        })
+      );
+      try {
+        await api.deleteRegistry(registry?.id).then(() => {
+          history.push(homeLinkPath || '/');
+          fetchRegistry();
+        });
+      } catch (error) {}
+    }
   };
 
   const onDeleteRegistry = () => {
@@ -112,51 +114,8 @@ export const ServiceRegistry: React.FC<ServiceRegistryProps> = ({
      */
   };
 
-  const renderWelcomeEmptyState = () => {
-    return (
-      <PageSection padding={{ default: 'noPadding' }} isFilled>
-        <MASEmptyState
-          emptyStateProps={{ variant: MASEmptyStateVariant.GettingStarted }}
-          titleProps={{ title: t('serviceRegistry.welcome_to_service_registry') }}
-          emptyStateBodyProps={{
-            body: t('serviceRegistry.welcome_empty_state_body'),
-          }}
-          buttonProps={{
-            title: t('serviceRegistry.create_service_registry'),
-            variant: ButtonVariant.primary,
-            onClick: createServiceRegistry,
-            isLoading: isLoading,
-            spinnerAriaValueText: isLoading ? t('common.loading') : undefined,
-          }}
-        />
-      </PageSection>
-    );
-  };
-
-  const renderUnauthorizedUserEmptyState = () => {
-    return (
-      <>
-        <ServiceRegistryHeader showKebab={false} />
-        <PageSection padding={{ default: 'noPadding' }} isFilled>
-          <MASEmptyState
-            emptyStateProps={{ variant: MASEmptyStateVariant.NoAccess }}
-            titleProps={{ title: t('serviceRegistry.unauthorized_empty_state_title') }}
-            emptyStateBodyProps={{
-              body: t('serviceRegistry.unauthorized_empty_state_body'),
-            }}
-            buttonProps={{
-              title: t('serviceRegistry.get_access_to_service_registry'),
-              variant: ButtonVariant.primary,
-              onClick: getAccessToServiceRegistry,
-            }}
-          />
-        </PageSection>
-      </>
-    );
-  };
-
   if (isUnauthorizedUser) {
-    return renderUnauthorizedUserEmptyState();
+    return <UnauthrizedUser getAccessToServiceRegistry={getAccessToServiceRegistry} />;
   }
 
   const renderView = () => {
@@ -168,18 +127,26 @@ export const ServiceRegistry: React.FC<ServiceRegistryProps> = ({
           notRequiredDrawerContentBackground={notRequiredDrawerContentBackground}
           onClose={onCloseDrawer}
         >
-          <FederatedModuleManager
+          {/* <FederatedModuleManager
             tenantId={tenantId}
             baseUIPath={baseUIPath}
             params={params}
             onConnectToRegistry={onConnectToRegistry}
             onDeleteRegistry={onDeleteRegistry}
             homeLinkPath={homeLinkPath}
+            activeFederatedModule={activeFederatedModule}
+          /> */}
+          <ServiceRegistryHeader
+            onConnectToRegistry={onConnectToRegistry}
+            onDeleteRegistry={onDeleteRegistry}
+            showBreadcrumb={showBreadcrumb}
+            activeBreadcrumbItemLabel={activeBreadcrumbItemLabel}
+            homeLinkPath={homeLinkPath}
           />
         </ServiceRegistryDrawer>
       );
     } else {
-      return renderWelcomeEmptyState();
+      return <WelcomeEmptyState createServiceRegistry={createServiceRegistry} isLoading={isLoading} />;
     }
   };
